+++ dlls/rpg.cpp
@@ -480,6 +480,12 @@ void CRpg::SecondaryAttack()
 
 void CRpg::WeaponIdle()
 {
+	// Reset when the player lets go of the trigger.
+	if ((m_pPlayer->pev->button & (IN_ATTACK | IN_ATTACK2)) == 0)
+	{
+		ResetEmptySound();
+	}
+
 	UpdateSpot();
 
 	if (m_flTimeWeaponIdle > UTIL_WeaponTimeBase())
@@ -508,7 +514,6 @@ void CRpg::WeaponIdle()
 			m_flTimeWeaponIdle = UTIL_WeaponTimeBase() + 6.1;
 		}
 
-		ResetEmptySound();
 		SendWeaponAnim(iAnim);
 	}
 	else
@@ -542,6 +547,16 @@ void CRpg::UpdateSpot()
 #endif
 }
 
+bool CRpg::IsUseable()
+{
+	// The client needs to fall through to WeaponIdle so check the ammo here.
+	if (m_pPlayer->ammo_rockets <= 0)
+	{
+		return false;
+	}
+
+	return CBasePlayerWeapon::IsUseable();
+}
 
 class CRpgAmmo : public CBasePlayerAmmo
 {